Finish mex command to pass to libs
